+++ cpu_usage_list.h
@@ -4,11 +4,11 @@
 typedef struct Cpu_usage_list Cpu_usage_list;
 
 Cpu_usage_list* cpu_usage_list_create(void);
-void cpu_usage_list_add(Cpu_usage_list* const list, const unsigned int cpu_num, const float percent);
+void cpu_usage_list_add(Cpu_usage_list* const list, const int cpu_num, const double percent);
 void cpu_usage_list_delete(Cpu_usage_list* list);
 size_t cpu_usage_list_get_size(Cpu_usage_list* list);
-unsigned int cpu_usage_list_get_cpu_num(Cpu_usage_list* list, int n);
-float cpu_usage_list_get_percent(Cpu_usage_list* list, int n);
+int cpu_usage_list_get_cpu_num(Cpu_usage_list* list, size_t n);
+double cpu_usage_list_get_percent(Cpu_usage_list* list, size_t n);
